Review feedback: rename package, move tests out of package, remove nested src/ dir in package
