+++ python/openai/tests/test_chat_completions.py
