+++ python/openai/tests/test_completions.py
