+++ python/openai/tests/vllm_models/llama-3.1-8b-instruct/1/model.json
